Show checklist once user submit a pull request. (#2194) (#2196)
This pull request makes sure that users are shown a checklist after
submitting a pull request to make sure a pull request meet some
criteria.

Co-authored-by: Yaara <3155222+letzya@users.noreply.github.com>

## Related Issue (if applicable)
<!-- If applicable, please reference to a related issue in this
repository. -->

## Description
<!-- Describe your changes in detail.
Please explain the change to the reviewer. For example, if it's a change
in the file name or directory
name please flag it out since it’s hard to compare text after such a
change -->
<!-- Provide a summary of your changes in the title above 
     Please remember to add the following: 
     - A description of the changes proposed in the pull request. 
Please explain the change to the reviewer. Pay special attention for
changes that are hard to spot,
for example, if it's a change in the file name or directory name please
flag it out since it’s hard
       to compare text after such a change 
- @mentions of the person or team responsible for reviewing proposed
changes. -->

## Motivation and Context
<!-- Why is this change required? What problem does it solve? -->

## Preview Link
<!-- Add a preview link to make it easier for reviewer to review. -->
## Screenshots (if appropriate)

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Fixing typo (please merge to production)
- [ ] Documenting a new feature (please merge to production)
- [ ] Documentation for future release (please do not merge to
production)
- [X] Something else (please add if needs merging to production or not)
merge from master

## Checklist
<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
- [X] I have [reviewed the guidelines](../CONTRIBUTING.md) for
contributing to this repository.
- [ ] I have [read the technical
guidelines](../CONTRIBUTING-TECHNICAL-GUIDE.md) for contributing to this
repository.
- [ ] Make sure you have started *your change* off *our latest
`master`*.
- [ ] I have added a preview link.

Co-authored-by: itachi sasuke <Keithwachira@gmail.com>